extracted out action buttons from cohortDetails
